Renamed abnormalType const filename
