+++ src/models/const/abnormal-type.js
